scripts/python/requirements.txt
boto3
python-dotenv
urllib3
pynacl